+++ interfaces.ts
@@ -15,22 +15,3 @@ export interface UserSample {
     };
   };
 }
-
-// For redux
-export interface UserState {
-  id: number;
-  name: string;
-  username: string;
-  email: string;
-  status: "idle" | "loading" | "succeeded" | "failed";
-  address: {
-    street: string;
-    suite: string;
-    city: string;
-    zipcode: string;
-    geo: {
-      lat: number;
-      lng: number;
-    };
-  };
-}
